flask_application/requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.18.0
annotated-types==0.5.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.6.2
boto3==1.26.76
botocore==1.29.76
cachetools==5.3.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
faiss-cpu==1.7.4
fastjsonschema==2.16.2
filelock==3.9.0
fire==0.5.0
Flask==2.3.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
InstructorEmbedding==1.0.1
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_fileid==0.7.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kiwisolver==1.4.4
langchain==0.0.270
langsmith==0.0.25
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.5
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.9
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
protobuf==3.20.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
Pygments==2.14.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.6
python-telegram-bot==20.1
pytz==2022.7.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.0
safetensors==0.3.2
sahi==0.11.11
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.20
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
ultralytics-yolov5==0.1.1
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.7
wincertstore==0.2
y-py==0.5.9
yarl==1.9.2
yolov5==7.0.9
ypy-websocket==0.8.2
zipp==3.14.0
